refactor: Move ToolState to dedicated tool_handler module
