+++ coauthor/agent/agent_base.py
@@ -27,7 +27,8 @@ from ..utils.xml import get_xml_format_from_files
 
 # Local imports - agent components
 from .agent_dataclass import AgentConfig, AgentSettings, AgentPrompts
-from .agent_state import AgentRoundState, AgentGlobalState, ToolState
+from .agent_state import AgentRoundState, AgentGlobalState
+from .tool_handler import ToolState
 from .logdb import create_log_entry, update_log_statistics, update_log_output_files
 from .model_base import ModelHandler
 from .output_handler import OutputHandler
