+++ coauthor/agent/agent_state.py
@@ -1,4 +1,4 @@
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 
 from ..logger import logger
 from .response_usage import OpenAIResponseUsage, AnthropicResponseUsage
@@ -43,34 +43,6 @@ class AgentRoundState:
             "output_file": self.output_file,
             "model_usage": self.model_usage,
         }
-
-
-@dataclass
-class ToolState:
-    """State for tool-specific runtime data that doesn't need to be logged."""
-
-    tex_count_stats: str | None = None  # Statistics about TeX document structure
-    first_k_tex_document: str | None = None  # First K lines of TeX document
-    last_response: str = ""  # Most recent model response
-    accumulated_output: str = ""  # Combined output from all responses
-    figure_files: list[str] = field(default_factory=list)  # Paths to figure files
-
-    @classmethod
-    def initialize(cls) -> "ToolState":
-        """Initialize a new ToolState object."""
-        return cls()
-
-    def update_last_response(self, response: str) -> None:
-        """Update the last response."""
-        self.last_response = response
-
-    def update_accumulated_output(self, output: str) -> None:
-        """Update the accumulated output."""
-        self.accumulated_output = output
-
-    def add_figure_files(self, files: list[str]) -> None:
-        """Add figure files to the list."""
-        self.figure_files.extend(files)
 
 
 @dataclass
